remote tf model moved to tf dir. services check backend for remote models
